Move fixtures out of tests due to class cannot be found
