+++ tests/Engines/ArrayEngineTest.php
@@ -11,9 +11,9 @@ use PHPUnit\Framework\TestCase;
 use stdClass;
 use Sti3bas\ScoutArray\ArrayStore;
 use Sti3bas\ScoutArray\Engines\ArrayEngine;
-use Sti3bas\ScoutArray\Tests\Fixtures\EmptySearchableModel;
-use Sti3bas\ScoutArray\Tests\Fixtures\SearchableModel;
-use Sti3bas\ScoutArray\Tests\Fixtures\SoftDeletableSearchableModel;
+use Sti3bas\ScoutArray\Fixtures\EmptySearchableModel;
+use Sti3bas\ScoutArray\Fixtures\SearchableModel;
+use Sti3bas\ScoutArray\Fixtures\SoftDeletableSearchableModel;
 
 class ArrayEngineTest extends TestCase
 {
